requirements/tests.txt
-r base.txt

coverage
ddt
django-nose
factory-boy
pylint-celery
pylint-django
pylint-plugin-utils
pytest-cov
pytest-django
pytest
